dashboard/requirements.txt
pyserial==3.5
pillow